Changed uint64 to uint32

+++ motor/mc_interface.h
@@ -66,7 +66,7 @@ float mc_interface_get_duty_cycle_now(void);
 float mc_interface_get_sampling_frequency_now(void);
 float mc_interface_get_rpm(void);
 float mc_interface_get_tacho_rpm(void); 
-uint64_t mc_interface_get_tacho_count(void);
+uint32_t mc_interface_get_tacho_count(void);
 float mc_interface_get_amp_hours(bool reset);
 float mc_interface_get_amp_hours_charged(bool reset);
 float mc_interface_get_watt_hours(bool reset);
